tests: moved modes tests to core/modes folder
