+++ tests/core/fixtures/compression/base.txt.bz2
