+++ tests/core/fixtures/compression/base.txt.gz
