+++ tests/core/fixtures/compression/base.txt.lzo
