+++ tests/core/fixtures/compression/base.txt.xz
