+++ tests/core/fixtures/compression/symbolic.gz
