+++ tests/core/fixtures/modes/copy_default.txt
