+++ tests/core/fixtures/modes/copy_full.txt
