+++ tests/core/fixtures/modes/flash_default.txt
