+++ tests/core/fixtures/modes/flash_full.txt
