+++ tests/core/fixtures/modes/imxkobs_default.txt
