+++ tests/core/fixtures/modes/imxkobs_full.txt
