+++ tests/core/fixtures/modes/raw_default.txt
