+++ tests/core/fixtures/modes/raw_full.txt
