+++ tests/core/fixtures/modes/tarball_default.txt
