+++ tests/core/fixtures/modes/tarball_full.txt
