+++ tests/core/fixtures/modes/ubifs_default.txt
