+++ tests/core/fixtures/modes/ubifs_full.txt
