+++ tests/core/modes/test_copy.py
