+++ tests/core/modes/test_flash.py
