+++ tests/core/modes/test_raw.py
