+++ tests/core/modes/test_tarball.py
